+++ src/helpers/get-user-id.ts
@@ -1,0 +1,3 @@
+import {IncomingMessage} from "http";
+
+export const getUserId = (req: IncomingMessage) => req.url?.split('/').pop();
