+++ src/routing/users/handlers/get-users-handler.ts
@@ -1,0 +1,8 @@
+import {handleResponse} from "../../../helpers/handle-response.js";
+import {IncomingMessage, ServerResponse} from "http";
+import {UsersRepository} from "../../../users-repository.js";
+
+export const getUsersHandler = (req: IncomingMessage, res: ServerResponse, usersRepository: UsersRepository) => {
+    handleResponse(res, usersRepository.getAll());
+    return usersRepository.getData();
+}
